Отдельный класс с main() для HttpTaskServer Удалено лишнее определение DateTimeFormatter Вместо исключений IOException при отправке ответов используется вывод сообщения в консоль.
