+++ src/dc/yandex/kanban/service/server/adapters/LocalDateTimeAdapter.java
@@ -10,12 +10,12 @@ import java.time.format.DateTimeFormatter;
 
 public class LocalDateTimeAdapter extends TypeAdapter<LocalDateTime> {
 
-    private static final DateTimeFormatter dtf = DateTimeFormatter.ofPattern("dd.MM.yyyy HH:mm");
+    public static final DateTimeFormatter defaultDateTimeFormatter = DateTimeFormatter.ISO_LOCAL_DATE_TIME;
 
     @Override
     public void write(final JsonWriter jsonWriter, final LocalDateTime localDateTime) throws IOException {
         if (localDateTime != null) {
-            jsonWriter.value(localDateTime.format(dtf));
+            jsonWriter.value(localDateTime.format(defaultDateTimeFormatter));
         } else {
             jsonWriter.nullValue();
         }
@@ -23,6 +23,6 @@ public class LocalDateTimeAdapter extends TypeAdapter<LocalDateTime> {
 
     @Override
     public LocalDateTime read(final JsonReader jsonReader) throws IOException {
-        return LocalDateTime.parse(jsonReader.nextString(), dtf);
+        return LocalDateTime.parse(jsonReader.nextString(), defaultDateTimeFormatter);
     }
 }
